refactor the way python is used in dune

See merge request extensions/dune-alugrid!113

Former-commit-id: 40b135d21c9c1156a8bbe32d9091159b5bdac5b8

+++ dune/alugrid/3d/indexsets.hh
@@ -501,7 +501,7 @@ namespace Dune {
       const int childOffSet = offset[creatorCodim][cd];
 
       alugrid_assert ( nChild < childOffSet );
-      alugrid_assert ( creatorNumber < nEntitiesFactor[creatorCodim] * nElements );
+      alugrid_assert ( (unsigned int)creatorNumber < nEntitiesFactor[creatorCodim] * nElements );
 
       typename IdType::IntegerType newChild =  creatorNumber * childOffSet  + nChild;
       for(int i=cd ; i > creatorCodim ; i--)
